+++ mne_bids/write.py
@@ -89,7 +89,8 @@ def _channels_tsv(raw, fname, overwrite=False, verbose=True):
                     ecg='ElectroCardioGram',
                     eog='ElectroOculoGram',
                     emg='ElectroMyoGram',
-                    misc='Miscellaneous')
+                    misc='Miscellaneous',
+                    bio='Biological')
     get_specific = ('mag', 'ref_meg', 'grad')
 
     # get the manufacturer from the file in the Raw object